Use a modern approach for cross-site request forgery protection
Take advantage of the Sec-Fetch-Site header, which is a fetch metadata
request header that modern browsers send, and that indicates whether a
request for a resource is coming from the same origin, the same site or
a different site. The server can use this to decide whether the request
should be allowed or not.

This approach is used by default, in a very conservative and strict way:
only requests with a Sec-Fetch-Site value of same-site or same-origin
are allowed. Requests that don't include a Sec-Fetch-Site header or that
have it set to none are rejected. Assuming a modern browser, user-initiated
requests (like direct navigation or clicking a bookmark) shouldn't cause
a non-GET/HEAD request, so these wouldn't be subject to the CSRF check
in any case. Requests that don't include the header could either be:
- Not from a browser, in which case they wouldn't include the
  authenticity token either. These would be excluded from the check in
  another way.
- From a browser, but having the header stripped by some middle proxy or
  firewal (we've seen this in the wild). In this case we can't allow the
  request without making the user vulnerable to CSRF.

For apps that can't use this approach yet because they need to support
older browsers that don't send the Sec-Fetch-Site header, the usual
authenticity token approach is kept as a fallback. When a decision can't
be clearly made with the header value because it's missing or 'none',
the token is checked as before.

The Origin check is not changed, it continues to work in the same way,
combined with the Sec-Fetch-Site or authenticity token checks.

References:
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Sec-Fetch-Site
- https://web.dev/articles/fetch-metadata
- https://github.com/golang/go/issues/73626
- https://www.alexedwards.net/blog/preventing-csrf-in-go